requirements.txt
Flask>=2.1.2
Flask-Login
Flask-WTF==0.15.1
Flask-SQLAlchemy>=2.5.1
flask-login
Flask-Bootstrap>=3.3.7.1
Flask-Nav>=0.6
Werkzeug>=2.0.3
pymysql==1.0.2
MarkupSafe==2.0.1
Jinja2==3.0.3
SQLAlchemy==1.3.19
email_validator==1.1.3
paho-mqtt